pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Test"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "asyncpg",
    "alembic",
    "pydantic[email]>=2.10.6",
    "python-multipart",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]",
    "fastapi-filter",
    "python-dateutil",
    "holidays",
    "typing-extensions",
    "pydantic-settings>=2.7.1",
    "us>=3.2.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
virtual = true
managed = true
dev-dependencies = ["ruff", "alembic"]


[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.ruff]
line-length = 79


